+++ src/components/dialog/dialog.stories.mdx
@@ -10,7 +10,7 @@ import Textbox from "../textbox";
 import {
   RadioButton,
   RadioButtonGroup,
-} from "../../__experimental__/components/radio-button";
+} from "../radio-button";
 import Fieldset from "../fieldset";
 import Loader from "../loader";
