+++ src/components/fieldset/fieldset-examples.stories.mdx
@@ -4,7 +4,7 @@ import Fieldset from "./";
 import {
   RadioButton,
   RadioButtonGroup,
-} from "../../__experimental__/components/radio-button";
+} from "../radio-button";
 import { Select, Option } from "../select";
 import Textbox from "../textbox";
