+++ src/components/fieldset/fieldset.stories.mdx
@@ -4,7 +4,7 @@ import Fieldset from "./";
 import {
   RadioButton,
   RadioButtonGroup,
-} from "../../__experimental__/components/radio-button";
+} from "../radio-button";
 import { Select, Option } from "../select";
 import StyledSystemProps from "../../../.storybook/utils/styled-system-props";
 import Textbox from "../textbox";
