+++ src/components/form/form.stories.mdx
@@ -7,7 +7,7 @@ import Textarea from "../textarea";
 import {
   RadioButton,
   RadioButtonGroup,
-} from "../../__experimental__/components/radio-button";
+} from "../radio-button";
 import DateInput from "../date";
 import { Select, Option } from "../select";
 import { Checkbox } from "../checkbox";
