+++ src/components/radio-button/radio-button-group.spec.js
@@ -4,11 +4,11 @@ import { mount } from "enzyme";
 import {
   assertStyleMatch,
   mockMatchMedia,
-} from "../../../__spec_helper__/test-utils";
+} from "../../__spec_helper__/test-utils";
 import { RadioButton, RadioButtonGroup } from ".";
 import RadioButtonGroupStyle from "./radio-button-group.style";
-import Fieldset from "../../../__internal__/fieldset";
-import Label from "../../../__internal__/label";
+import Fieldset from "../../__internal__/fieldset";
+import Label from "../../__internal__/label";
 
 const buttonValues = ["test-1", "test-2"];
 const name = "test-group";
