+++ src/components/radio-button/radio-button.style.js
@@ -1,13 +1,13 @@
 import styled, { css } from "styled-components";
 import { margin } from "styled-system";
-import FieldHelpStyle from "../../../__internal__/field-help/field-help.style";
-import CheckboxStyle from "../../../components/checkbox/checkbox.style";
-import HiddenCheckableInputStyle from "../../../__internal__/checkable-input/hidden-checkable-input.style";
-import { StyledCheckableInput } from "../../../__internal__/checkable-input/checkable-input.style";
-import StyledCheckableInputSvgWrapper from "../../../__internal__/checkable-input/checkable-input-svg-wrapper.style";
-import { StyledLabelContainer } from "../../../__internal__/label/label.style";
-import baseTheme from "../../../style/themes/base";
-import FormFieldStyle from "../../../__internal__/form-field/form-field.style";
+import FieldHelpStyle from "../../__internal__/field-help/field-help.style";
+import CheckboxStyle from "../checkbox/checkbox.style";
+import HiddenCheckableInputStyle from "../../__internal__/checkable-input/hidden-checkable-input.style";
+import { StyledCheckableInput } from "../../__internal__/checkable-input/checkable-input.style";
+import StyledCheckableInputSvgWrapper from "../../__internal__/checkable-input/checkable-input-svg-wrapper.style";
+import { StyledLabelContainer } from "../../__internal__/label/label.style";
+import baseTheme from "../../style/themes/base";
+import FormFieldStyle from "../../__internal__/form-field/form-field.style";
 
 const RadioButtonStyle = styled(CheckboxStyle)`
   ${({ disabled, fieldHelpInline, reverse, size, theme, inline }) => css`
